refactor Renamed and moved files to offline. Deleted old.
