--- mule/offline3_train_model_OBSELETE.py
@@ -1,535 +1,0 @@
-#import sys
-import glob,os
-#import json
-import pandas as pd
-#import tensorflow as tf
-import logging
-#import zipfile
-#import re
-#import datetime
-#import cv2
-import h5py
-
-from tensorflow.python.keras.layers import Input
-from tensorflow.python.keras.models import Model, load_model
-from tensorflow.python.keras.layers import Convolution2D
-from tensorflow.python.keras.layers import Dropout, Flatten, Dense, Cropping2D, Lambda
-from tensorflow.python.keras.callbacks import ModelCheckpoint, EarlyStopping
-from tensorflow.python.keras import backend as K
-import tensorflow as tf
-# Check versions
-assert tf.__version__ == '1.8.0'
-
-
-#%% Constants
-BATCH_SIZE = 128
-TRAIN_TEST_SPLIT = 0.8
-MODEL_NAME = 'Test1'
-SIZE_H = 120
-SIZE_W = 160
-
-#%% IO
-# Project
-LOCAL_PROJECT_PATH = glob.glob(os.path.expanduser('~/MULE DATA'))[0]
-assert os.path.exists(LOCAL_PROJECT_PATH)
-
-# Model path
-new_model_dir = os.path.join(selected_data['this_dir'],'model')
-if not os.path.exists(new_model_dir):
-    os.mkdir(new_model_dir)
-fullpath_new_model = os.path.join(new_model_dir,'{}.hd5'.format(MODEL_NAME))
-
-#%%
-
-# =============================================================================
-# Training Generator
-# =============================================================================
-train_datagen = tf.keras.preprocessing.image.ImageDataGenerator(
-        #shear_range=0.2,
-        #zoom_range=0.5,
-        #width_shift_range=0.5,
-        #height_shift_range=?,
-        #rotation_range = 10,
-        #horizontal_flip=True,
-        #fill_mode = 'constant',
-        #cval = 0., # value to fill input images when fill_mode='constant'
-        #label_cval = 11. # value to fill labels when fill_mode='constant'
-        )
-
-#train_flow = train_datagen.flow_from_directory(
-#        X_path, Y_path,
-#        target_size=(SIZE_H, SIZE_W),
-#        batch_size=batch_size,
-#        shuffle = True,
-#        #save_to_dir = os.path.join('camvid', 'debugs'), # uncomment to save (image, label) to dir for debuging mode
-#        #save_prefix = 'd',
-#        #save_format = 'png',
-#        class_mode=None
-#        )
-
-#flow_gen = train_datagen.flow(X_train,y)
-
-
-
-#%%
-# =============================================================================
-# Define the models
-# =============================================================================
-
-#%%
-""""
-
-keras.py
-
-functions to run and train autopilots using keras
-
-"""
-
-
-#from donkeycar import util
-
-
-class KerasPilot:
-
-    def load(self, model_path):
-        self.model = load_model(model_path)
-
-    def shutdown(self):
-        pass
-
-    def train(self, train_gen, val_gen,
-              saved_model_path, epochs=100, steps=100, train_split=0.8,
-              verbose=1, min_delta=.0005, patience=5, use_early_stop=True):
-        """
-        train_gen: generator that yields an array of images an array of
-
-        """
-
-        # checkpoint to save model after each epoch
-        save_best = ModelCheckpoint(saved_model_path,
-                                    monitor='val_loss',
-                                    verbose=verbose,
-                                    save_best_only=True,
-                                    mode='min')
-
-        # stop training if the validation error stops improving.
-        early_stop = EarlyStopping(monitor='val_loss',
-                                   min_delta=min_delta,
-                                   patience=patience,
-                                   verbose=verbose,
-                                   mode='auto')
-
-        callbacks_list = [save_best]
-
-        if use_early_stop:
-            callbacks_list.append(early_stop)
-
-        hist = self.model.fit_generator(
-            train_gen,
-            steps_per_epoch=steps,
-            epochs=epochs,
-            verbose=1,
-            validation_data=val_gen,
-            callbacks=callbacks_list,
-            validation_steps=steps * (1.0 - train_split) / train_split)
-        return hist
-
-
-class KerasCategorical(KerasPilot):
-    def __init__(self, model=None, *args, **kwargs):
-        super(KerasCategorical, self).__init__(*args, **kwargs)
-        if model:
-            self.model = model
-        else:
-            self.model = default_categorical()
-
-    def run(self, img_arr):
-        img_arr = img_arr.reshape((1,) + img_arr.shape)
-        angle_binned, throttle = self.model.predict(img_arr)
-        angle_unbinned = util.data.linear_unbin(angle_binned[0])
-        return angle_unbinned, throttle[0][0]
-
-
-class KerasLinear(KerasPilot):
-    def __init__(self, model=None, num_outputs=None, *args, **kwargs):
-        super(KerasLinear, self).__init__(*args, **kwargs)
-        if model:
-            self.model = model
-        elif num_outputs is not None:
-            self.model = default_n_linear(num_outputs)
-        else:
-            self.model = default_linear()
-
-    def run(self, img_arr):
-        img_arr = img_arr.reshape((1,) + img_arr.shape)
-        outputs = self.model.predict(img_arr)
-        # print(len(outputs), outputs)
-        steering = outputs[0]
-        throttle = outputs[1]
-        return steering[0][0], throttle[0][0]
-
-
-
-
-
-class KerasLinear(KerasPilot):
-    def __init__(self, model=None, num_outputs=None, *args, **kwargs):
-        super(KerasLinear, self).__init__(*args, **kwargs)
-        if model:
-            self.model = model
-        elif num_outputs is not None:
-            self.model = default_n_linear(num_outputs)
-        else:
-            self.model = default_linear()
-
-    def run(self, img_arr):
-        img_arr = img_arr.reshape((1,) + img_arr.shape)
-        outputs = self.model.predict(img_arr)
-        # print(len(outputs), outputs)
-        steering = outputs[0]
-        throttle = outputs[1]
-        return steering[0][0], throttle[0][0]
-
-
-
-
-
-def default_linear():
-    img_in = Input(shape=(120, 160, 3), name='img_in')
-    x = img_in
-    x = Convolution2D(24, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(32, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)
-
-    x = Flatten(name='flattened')(x)
-    x = Dense(100, activation='linear')(x)
-    x = Dropout(.1)(x)
-    x = Dense(50, activation='linear')(x)
-    x = Dropout(.1)(x)
-    # categorical output of the angle
-    angle_out = Dense(1, activation='linear', name='angle_out')(x)
-
-    # continous output of throttle
-    throttle_out = Dense(1, activation='linear', name='throttle_out')(x)
-
-    model = Model(inputs=[img_in], outputs=[angle_out, throttle_out])
-
-    model.compile(optimizer='adam',
-                  loss={'angle_out': 'mean_squared_error',
-                        'throttle_out': 'mean_squared_error'},
-                  loss_weights={'angle_out': 0.5, 'throttle_out': .5})
-
-    return model
-
-#%%
-# ******************************************************************************************************
-#FROM DK: 
-#
-
-# python ~/mycar/manage.py train --tub <tub folder names comma separated> --model ./models/mypilot
-
-def train(cfg, tub_names, new_model_path, base_model_path=None ):
-    """
-    use the specified data in tub_names to train an artifical neural network
-    saves the output trained model as model_name
-    """
-    X_keys = ['cam/image_array']
-    y_keys = ['user/angle', 'user/throttle']
-    def train_record_transform(record):
-        """ convert categorical steering to linear and apply image augmentations """
-        record['user/angle'] = dk.util.data.linear_bin(record['user/angle'])
-        # TODO add augmentation that doesn't use opencv
-        return record
-
-    def val_record_transform(record):
-        """ convert categorical steering to linear """
-        record['user/angle'] = dk.util.data.linear_bin(record['user/angle'])
-        return record
-
-    new_model_path = os.path.expanduser(new_model_path)
-
-    kl = KerasCategorical()
-    if base_model_path is not None:
-        base_model_path = os.path.expanduser(base_model_path)
-        kl.load(base_model_path)
-
-    print('tub_names', tub_names)
-    if not tub_names:
-        tub_names = os.path.join(cfg.DATA_PATH, '*')
-    tubgroup = TubGroup(tub_names)
-    train_gen, val_gen = tubgroup.get_train_val_gen(X_keys, y_keys,
-                                                    train_record_transform=train_record_transform,
-                                                    val_record_transform=val_record_transform,
-                                                    batch_size=cfg.BATCH_SIZE,
-                                                    train_frac=cfg.TRAIN_TEST_SPLIT)
-
-    total_records = len(tubgroup.df)
-    total_train = int(total_records * cfg.TRAIN_TEST_SPLIT)
-    total_val = total_records - total_train
-    print('train: %d, validation: %d' % (total_train, total_val))
-    steps_per_epoch = total_train // cfg.BATCH_SIZE
-    print('steps_per_epoch', steps_per_epoch)
-
-    kl.train(train_gen,
-             val_gen,
-             saved_model_path=new_model_path,
-             steps=steps_per_epoch,
-             train_split=cfg.TRAIN_TEST_SPLIT)
-
-
-
-
-
-
-
-
-#%%
-def default_categorical():
-    img_in = Input(shape=(120, 160, 3),
-                   name='img_in')  # First layer, input layer, Shape comes from camera.py resolution, RGB
-    x = img_in
-    x = Convolution2D(24, (5, 5), strides=(2, 2), activation='relu')(x)  # 24 features, 5 pixel x 5 pixel kernel (convolution, feauture) window, 2wx2h stride, relu activation
-    x = Convolution2D(32, (5, 5), strides=(2, 2), activation='relu')(x)  # 32 features, 5px5p kernel window, 2wx2h stride, relu activatiion
-    x = Convolution2D(64, (5, 5), strides=(2, 2), activation='relu')(x)  # 64 features, 5px5p kernal window, 2wx2h stride, relu
-    x = Convolution2D(64, (3, 3), strides=(2, 2), activation='relu')(x)  # 64 features, 3px3p kernal window, 2wx2h stride, relu
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)  # 64 features, 3px3p kernal window, 1wx1h stride, relu
-
-    # Possibly add MaxPooling (will make it less sensitive to position in image).  Camera angle fixed, so may not to be needed
-
-    x = Flatten(name='flattened')(x)  # Flatten to 1D (Fully connected)
-    x = Dense(100, activation='relu')(x)  # Classify the data into 100 features, make all negatives 0
-    x = Dropout(.1)(x)  # Randomly drop out (turn off) 10% of the neurons (Prevent overfitting)
-    x = Dense(50, activation='relu')(x)  # Classify the data into 50 features, make all negatives 0
-    x = Dropout(.1)(x)  # Randomly drop out 10% of the neurons (Prevent overfitting)
-    # categorical output of the angle
-    angle_out = Dense(15, activation='softmax', name='angle_out')(
-        x)  # Connect every input with every output and output 15 hidden units. Use Softmax to give percentage. 15 categories and find best one based off percentage 0.0-1.0
-
-    # continous output of throttle
-    throttle_out = Dense(1, activation='relu', name='throttle_out')(x)  # Reduce to 1 number, Positive number only
-
-    model = Model(inputs=[img_in], outputs=[angle_out, throttle_out])
-    model.compile(optimizer='adam',
-                  loss={'angle_out': 'categorical_crossentropy',
-                        'throttle_out': 'mean_absolute_error'},
-                  loss_weights={'angle_out': 0.9, 'throttle_out': .01})
-
-    return model
-
-
-
-
-def linear_unbin_layer(tnsr):
-    bin = K.constant((2 / 14), dtype='float32')
-    norm = K.constant(1, dtype='float32')
-
-    b = K.cast(K.argmax(tnsr), dtype='float32')
-    a = b - norm
-    # print('linear_unbin_layer out: {}'.format(a))
-    return a
-
-
-def default_catlin():
-    """
-    Categorial Steering output before linear conversion.
-    :return:
-    """
-    img_in = Input(shape=(120, 160, 3),
-                   name='img_in')  # First layer, input layer, Shape comes from camera.py resolution, RGB
-    x = img_in
-    x = Convolution2D(24, (5, 5), strides=(2, 2), activation='relu')(x)  # 24 features, 5 pixel x 5 pixel kernel (convolution, feauture) window, 2wx2h stride, relu activation
-    x = Convolution2D(32, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (5, 5), strides=(2, 2), activation='relu')(x)  # 64 features, 5px5p kernal window, 2wx2h stride, relu
-    x = Convolution2D(64, (3, 3), strides=(2, 2), activation='relu')(x)  # 64 features, 3px3p kernal window, 2wx2h stride, relu
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)  # 64 features, 3px3p kernal window, 1wx1h stride, relu
-
-    # Possibly add MaxPooling (will make it less sensitive to position in image).  Camera angle fixed, so may not to be needed
-
-    x = Flatten(name='flattened')(x)  # Flatten to 1D (Fully connected)
-    x = Dense(100, activation='relu')(x)  # Classify the data into 100 features, make all negatives 0
-    x = Dropout(.1)(x)  # Randomly drop out (turn off) 10% of the neurons (Prevent overfitting)
-    x = Dense(50, activation='relu')(x)  # Classify the data into 50 features, make all negatives 0
-    x = Dropout(.1)(x)  # Randomly drop out 10% of the neurons (Prevent overfitting)
-    # categorical output of the angle
-    angle_cat_out = Dense(15, activation='softmax', name='angle_cat_out')(x)
-    angle_out = Dense(1, activation='sigmoid', name='angle_out')(angle_cat_out)
-    # angle_out = Lambda(linear_unbin_layer, output_shape=(1,1, ), name='angle_out')(angle_cat_out)
-
-    # continuous output of throttle
-    throttle_out = Dense(1, activation='relu', name='throttle_out')(x)  # Reduce to 1 number, Positive number only
-
-    model = Model(inputs=[img_in], outputs=[angle_out, throttle_out])
-    model.compile(optimizer='adam',
-                  loss={'angle_out': 'mean_squared_error',
-                        'throttle_out': 'mean_absolute_error'},
-                  loss_weights={'angle_out': 0.9, 'throttle_out': .01})
-
-    return model
-
-
-def default_linear():
-    img_in = Input(shape=(120, 160, 3), name='img_in')
-    x = img_in
-    x = Convolution2D(24, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(32, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)
-
-    x = Flatten(name='flattened')(x)
-    x = Dense(100, activation='linear')(x)
-    x = Dropout(.1)(x)
-    x = Dense(50, activation='linear')(x)
-    x = Dropout(.1)(x)
-    # categorical output of the angle
-    angle_out = Dense(1, activation='linear', name='angle_out')(x)
-
-    # continous output of throttle
-    throttle_out = Dense(1, activation='linear', name='throttle_out')(x)
-
-    model = Model(inputs=[img_in], outputs=[angle_out, throttle_out])
-
-    model.compile(optimizer='adam',
-                  loss={'angle_out': 'mean_squared_error',
-                        'throttle_out': 'mean_squared_error'},
-                  loss_weights={'angle_out': 0.5, 'throttle_out': .5})
-
-    return model
-
-
-def default_n_linear(num_outputs):
-    img_in = Input(shape=(120, 160, 3), name='img_in')
-    x = img_in
-    x = Cropping2D(cropping=((60, 0), (0, 0)))(x)  # trim 60 pixels off top
-    x = Lambda(lambda x: x / 127.5 - 1.)(x)  # normalize and re-center
-    x = Convolution2D(24, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(32, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (5, 5), strides=(1, 1), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)
-
-    x = Flatten(name='flattened')(x)
-    x = Dense(100, activation='relu')(x)
-    x = Dropout(.1)(x)
-    x = Dense(50, activation='relu')(x)
-    x = Dropout(.1)(x)
-
-    outputs = []
-
-    for i in range(num_outputs):
-        outputs.append(Dense(1, activation='linear', name='n_outputs' + str(i))(x))
-
-    model = Model(inputs=[img_in], outputs=outputs)
-
-    model.compile(optimizer='adam',
-                  loss='mse')
-
-    return model
-
-
-#%% My Linear
-
-class KerasLinearMJ(KerasPilot):
-    def __init__(self, model=None, num_outputs=None, *args, **kwargs):
-        super(KerasLinearMJ, self).__init__(*args, **kwargs)
-        self.model = default_linearMJ()
-
-    def run(self, img_arr):
-        img_arr = img_arr.reshape((1,) + img_arr.shape)
-        outputs = self.model.predict(img_arr)
-        # print(len(outputs), outputs)
-        steering = outputs[0]
-        throttle = outputs[1]
-        return steering[0][0], throttle[0][0]
-
-    
-def default_linearMJ():
-    img_in = Input(shape=(120, 160, 3), name='img_in')
-    x = img_in
-    x = Convolution2D(24, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(32, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (5, 5), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(2, 2), activation='relu')(x)
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)
-
-    x = Flatten(name='flattened')(x)
-    x = Dense(100, activation='linear')(x)
-    x = Dropout(.1)(x)
-    x = Dense(50, activation='linear')(x)
-    x = Dropout(.1)(x)
-    # categorical output of the angle
-    angle_out = Dense(1, activation='linear', name='angle_out')(x)
-
-    # continous output of throttle
-    #model = Model(inputs=[img_in], outputs=[angle_out])
-    model = Model(inputs=img_in, outputs=angle_out)
-
-    model.compile(optimizer='adam',
-                  loss={'angle_out': 'mean_squared_error'})
-
-    return model
-
-
-#%% My categorical
-def my_default_categorical():
-    img_in = Input(shape=(120, 160, 3),
-                   name='img_in')  # First layer, input layer, Shape comes from camera.py resolution, RGB
-    x = img_in
-    x = Convolution2D(24, (5, 5), strides=(2, 2), activation='relu')(x)  # 24 features, 5 pixel x 5 pixel kernel (convolution, feauture) window, 2wx2h stride, relu activation
-    x = Convolution2D(32, (5, 5), strides=(2, 2), activation='relu')(x)  # 32 features, 5px5p kernel window, 2wx2h stride, relu activatiion
-    x = Convolution2D(64, (5, 5), strides=(2, 2), activation='relu')(x)  # 64 features, 5px5p kernal window, 2wx2h stride, relu
-    x = Convolution2D(64, (3, 3), strides=(2, 2), activation='relu')(x)  # 64 features, 3px3p kernal window, 2wx2h stride, relu
-    x = Convolution2D(64, (3, 3), strides=(1, 1), activation='relu')(x)  # 64 features, 3px3p kernal window, 1wx1h stride, relu
-
-    # Possibly add MaxPooling (will make it less sensitive to position in image).  Camera angle fixed, so may not to be needed
-
-    x = Flatten(name='flattened')(x)  # Flatten to 1D (Fully connected)
-    x = Dense(100, activation='relu')(x)  # Classify the data into 100 features, make all negatives 0
-    x = Dropout(.1)(x)  # Randomly drop out (turn off) 10% of the neurons (Prevent overfitting)
-    x = Dense(50, activation='relu')(x)  # Classify the data into 50 features, make all negatives 0
-    x = Dropout(.1)(x)  # Randomly drop out 10% of the neurons (Prevent overfitting)
-    # categorical output of the angle
-    angle_out = Dense(15, activation='softmax', name='angle_out')(
-        x)  # Connect every input with every output and output 15 hidden units. Use Softmax to give percentage. 15 categories and find best one based off percentage 0.0-1.0
-
-    # continous output of throttle
-    #throttle_out = Dense(1, activation='relu', name='throttle_out')(x)  # Reduce to 1 number, Positive number only
-
-    model = Model(inputs=img_in, outputs=angle_out)
-    model.compile(optimizer='adam',
-                  loss={'angle_out': 'categorical_crossentropy'})
-
-    return model
-#%%
-this_model = my_default_categorical()
-flowgen = train_datagen.flow(X_train[0:20],binned_y_matrix,batch_size=1)
-x_1, y_1 = flowgen.next()
-
-flowgen = train_datagen.flow(X_train[0:20],binned_y_matrix[0:20],batch_size=64)
-this_model.fit_generator(flowgen,epochs=20)
-
-
-#%%
-# =============================================================================
-# Get the CNN
-# =============================================================================
-kl = KerasLinearMJ()
-kl.model.summary()
-
-# Test the generator
-flowgen = train_datagen.flow(X_train,y,batch_size=1)
-x_1, y_1 = flowgen.next()
-
-#%%
-flowgen = train_datagen.flow(X_train,y,batch_size=32)
-kl.model.fit_generator(flowgen,epochs=10)
-#kl.model.fit_generator(flow_gen)
-
-
-
-#%%
-
-kl.model.save(fullpath_new_model)
-
